fix: migrate sending emails logic to Mail.store
